Add QMP commands matching the functionality.
Paused jobs cannot be canceled without first resuming them.  This
ensures that I/O errors are never missed by management.  However, an
optional force argument can be specified to allow that.  Right now,
jobs do not see the difference between a forced and a normal resume.
This can be changed in the future if needed.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>